themingengine: change coordinates for _gtk_theming_engine_paint_spinner
Instead of assuming the current point is the center of the spinner and
rotating around it, make it behave like gtk_render_icon(), i.e. the
current point is at the top-left corner of the desired spinner square.

This is needed for subsequent commits that add blur support to spinner
shadows.