Implement _gtk_clipboard_store_all()
This pushes the clipboard contents to the OS X clipboard when the
application is quit. Without doing this, clipboard data set by a GTK+
application cannot be accessed after the clipboard has been quit.

Currently, we implement this the easy way because the clipboard
support is fully implemented in GTK+. In the future this might change.