Visual C++ projects: Update .dll/.lib naming
Remove the "-win32-" from the output file names for the GDK and GTK+ DLLs,
like what is now done for quite a while on other platforms
(and MinGW builds), for consistency reasons.  This is due to GDK/GTK+
are buildable with multiple backends.

Note: For references, the Windows build only builds the Win32 backend
for the time being.